organized some things
