+++ lib/main.dart
@@ -10,7 +10,7 @@ import 'package:gthr/screens/HomePage/homepage.dart';
 import 'package:gthr/screens/ProfileScreen/profile_main.dart';
 import 'package:gthr/screens/Settings/settings.dart';
 import 'package:gthr/services/auth.dart';
-import 'package:gthr/splash_screen.dart';
+import 'package:gthr/shared/splash_screen.dart';
 import 'package:provider/provider.dart';
 
 void main() async {
